# public/swagger.json
{
	"openapi": "3.0.0",
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"ConsoleDTO": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"name": {
						"type": "string"
					},
					"manufacturer": {
						"type": "string"
					}
				},
				"required": [
					"name",
					"manufacturer"
				],
				"type": "object",
				"additionalProperties": false
			},
			"GameDTO": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"title": {
						"type": "string"
					},
					"console": {
						"$ref": "#/components/schemas/ConsoleDTO"
					}
				},
				"required": [
					"title"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CreateGameDTO": {
				"properties": {
					"title": {
						"type": "string"
					},
					"console": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"title",
					"console"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UpdateGameDTO": {
				"properties": {
					"title": {
						"type": "string"
					},
					"console": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"title",
					"console"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ReviewDTO": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"game": {
						"$ref": "#/components/schemas/GameDTO"
					},
					"rating": {
						"type": "number",
						"format": "double"
					},
					"reviewText": {
						"type": "string"
					}
				},
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "api-review-game",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"paths": {
		"/reviews/{id}": {
			"delete": {
				"operationId": "DeleteReview",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"tags": [
					"Reviews"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/games": {
			"get": {
				"operationId": "GetAllGames",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/GameDTO"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Games"
				],
				"security": [],
				"parameters": []
			},
			"post": {
				"operationId": "AddGame",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/GameDTO"
								}
							}
						}
					}
				},
				"tags": [
					"Games"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateGameDTO"
							}
						}
					}
				}
			}
		},
		"/games/{id}": {
			"get": {
				"operationId": "GetConsoleById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/GameDTO"
										}
									],
									"nullable": true
								}
							}
						}
					}
				},
				"tags": [
					"Games"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"patch": {
				"operationId": "UpdateGame",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/GameDTO"
										}
									],
									"nullable": true
								}
							}
						}
					}
				},
				"tags": [
					"Games"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateGameDTO"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteGame",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"tags": [
					"Games"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/games/{id}/reviews": {
			"get": {
				"operationId": "GetConsoleGames",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/ReviewDTO"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Games"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/consoles": {
			"get": {
				"operationId": "GetAllConsole",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/ConsoleDTO"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Consoles"
				],
				"security": [],
				"parameters": []
			},
			"post": {
				"operationId": "CreateConsole",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ConsoleDTO"
								}
							}
						}
					}
				},
				"tags": [
					"Consoles"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ConsoleDTO"
							}
						}
					}
				}
			}
		},
		"/consoles/{id}": {
			"get": {
				"operationId": "GetConsoleById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/ConsoleDTO"
										}
									],
									"nullable": true
								}
							}
						}
					}
				},
				"tags": [
					"Consoles"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"delete": {
				"operationId": "DeleteConsole",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"tags": [
					"Consoles"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"patch": {
				"operationId": "UpdateConsole",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/ConsoleDTO"
										}
									],
									"nullable": true
								}
							}
						}
					}
				},
				"tags": [
					"Consoles"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ConsoleDTO"
							}
						}
					}
				}
			}
		},
		"/consoles/{id}/games": {
			"get": {
				"operationId": "GetConsoleGames",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/GameDTO"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Consoles"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}